+++ @xen-orchestra/lite/src/components/host/network/HostPifsTable.vue
